15.07.2023 - МОДУЛИ js решил домашку!
