+++ index.html
@@ -18,5 +18,5 @@
       </div>
     </div>
   </body>
-  <script src="index.js"></script>
+  <script type="module" src="index.js"></script>
 </html>
